requirements.txt
tensorflow==2.17.1
keras==3.5.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.2
requests==2.28.2
gdown==4.5.2
python-dotenv==1.0.0
jupyter==1.0.0
ipython==8.12.3
flask==2.2.5
werkzeug==2.2.3
gunicorn==20.1.0
pillow==9.5.0
flask-cors
kagglehub
